+++ app/assets/stylesheets/color.scss
@@ -1,1 +1,2 @@
 $bg-logo: #0053ab;
+$red-color: red;
