+++ app/controllers/admin/admin_controller.rb
@@ -1,0 +1,3 @@
+class Admin::AdminController < ApplicationController
+  layout "admin/layouts/application"
+end
