+++ app/controllers/admin/admin_static_pages_controller.rb
@@ -1,0 +1,5 @@
+class Admin::AdminStaticPagesController < Admin::AdminController
+  before_action :check_login?
+
+  def home; end
+end
