--- app/controllers/products_controller.rb
@@ -1,5 +1,0 @@
-class ProductsController < ApplicationController
-  def index
-    @pagy, @products = pagy Product.all, items: Settings.product.items
-  end
-end
