+++ app/views/admin/admin_static_pages/home.html.erb
