+++ app/views/admin/layouts/application.html.erb
@@ -5,13 +5,11 @@
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-
     <%= stylesheet_link_tag "application", media: "all", "data-turbolinks-track": "reload" %>
     <%= javascript_pack_tag "application", "data-turbolinks-track": "reload" %>
   </head>
 
   <body class="bg-light">
-    <!-- TODO -->
-    <%= render "sessions/login_page" %>
+    <%= render "admin/layouts/admin_page" %>
   </body>
 </html>
