+++ app/views/admin/products/_product.html.erb
@@ -2,7 +2,11 @@
   <td><%= product.id %></td>
   <td><%= product.name %></td>
   <td><%= product.description %></td>
-  <td><img src="product.image" class="img-150"></td>
+  <td>
+    <% if product.image.attached? %>
+    <%= image_tag product.image, class: "img-150" %>
+    <% end %>
+  </td>
   <td><%= product.unit_price %> <%= t ".denominations" %></td>
   <td><%= product.category.name %></td>
   <td><input type="checkbox" <% if product.is_discontinued %>checked<% end %> disabled></td>
