+++ app/views/admin/products/index.html.erb
@@ -1,11 +1,11 @@
 <% provide :title, t(".title_page") %>
-<div class="row my-5 mx-4">
+<div class="row mt-2 mb-5 mx-4">
   <div class="col-md-12">
     <div class="card shadow">
       <div class="card-header d-flex justify-content-between align-center">
         <h1><%= t ".title" %></h1>
         <div>
-          <%= link_to t(".btn_add"), "#", class: "btn btn-success" %>
+          <%= link_to t(".btn_add"), new_admin_product_path, class: "btn btn-success" %>
         </div>
       </div>
       <div class="card-body pt-0 pb-2">
