+++ app/views/admin/products/new.html.erb
@@ -1,0 +1,13 @@
+<% provide :title, t(".title_page") %>
+<div class="row mt-2 mb-5 mx-4">
+  <div class="col-md-12">
+    <div class="card shadow">
+      <div class="card-header">
+        <h1><%= t ".title" %></h1>
+      </div>
+      <div class="card-body">
+        <%= render "form" %>
+      </div>
+    </div>
+  </div>
+</div>
