+++ app/views/sessions/new.html.erb
@@ -1,6 +1,11 @@
 <!-- TODO -->
 <% provide :title, t(".title") %>
 <div class="row">
+  <div class="col-8 mx-auto mb-3">
+    <% flash.each do |message_type, message| %>
+      <div class="alert alert-<%= message_type %>" role="alert"><%= message %></div>
+    <% end %>
+  </div>
   <div class="col-8 mx-auto">
     <div class="card">
       <div class="card-header">
